requirements.txt
Flask==3.0.3
langchain==0.2.1
python-dotenv==1.0.1
google-generativeai==0.5.4
langchain-google-genai==1.0.5
pytest==8.2.1